Add the mw-body-content class to the <article> element for consistency with other skins
MonoBook and Vector have this class on the #bodyContent div as well as the #siteNotice div, so I'm not sure if this should be placed on
.liberty-content-main instead of the <article> element, but placing it on
the <article> seems more logical to me.

This class is needed and used by the core ResourceLoader module
'mediawiki.skinning.content.externallinks' which provides external links
icons, so should Liberty ever want to make use of it, having this class on
some container element is a prerequisite.