Allow turning the live recent changes feature off
Also added i18n for live-recent.js...which doesn't work for some odd
reason. There are no functional changes in that JS file as the only
"functional" change is commented out; the diff is messy because I
reindented the code a bit.

Furthermore declared live-recent.js's dependency on core mediawiki.util
module in skin.json to ensure that mw.util is always defined.