Upgrade Bootstrap to version 3.3.7 to stop JavaScript errors in MW 1.32+ due to "too new" jQuery bundled with MW
Also stop loading a custom, outdated jQuery in the skins.liberty.bootstrap ResourceLoader module.
MediaWiki has loaded jQuery for quite many versions by default on *all* page loads.

In MediaWiki 1.32+ the version of jQuery bundled with MediaWiki and thus loaded automatically on all page loads is "too new" for Bootstrap v.3.3.6.
To play it safe, I decided to upgrade Bootstrap only to v.3.3.7 because it's known to work with the version of jQuery that MW 1.32+ ships with (see https://github.com/twbs/bootstrap/issues/16834) even though the current release of Bootstrap is v.4.3.1. I've tested 3.3.7 locally and confirmed that it's working, e.g. the user menu shows again when you click on it and so on.

Added the unminified, single-file bootstrap.js version of Bootstrap v.3.3.7 to the repository as well, even though it's not used, but having already minified source files without the unminified "real" source in the repository is kinda annoying, and from a security point of view, you'll wanna make sure that the source is there.

Bootstrap files downloaded from GitHub (bootstrap-3.3.7-dist.zip).