Awful, quick fix for the new RC Filters in MW 1.32+
This really isn't the correct way to do this, but this works...

With the new, OOUI-ful RCFilter stuff, this rule causes the "hamburger" icon on the "Filter changes (use menu or search for filter)" <input> area to be mispositioned, and additionally something here makes that input one or two pixels too tall so it "overflows". Adding a new :not() selector here fixes both issues, but it's not the clean, nice, scalable solution I'd want.