Fix some special page names' capitalization
Passing an incorrectly capitalized name to SpecialPage#getTitleFor causes a notice when development notices are enabled. For legacy reasons, older special pages' names (e.g. Userlogin, Whatlinkshere) are expected to be in lowercase except for the first letter, but newer special pages are assumed to be CamelCased (e.g. CreateAccount, EditTags, etc.).